recupera migrations deletadas e finaliza course model

+++ backend/app/Models/Course.php
@@ -17,24 +17,40 @@ class Course extends Model
 
     public function saveNewCourse($newCourse){
         $course = new Course();
-        $course = Course::fillCourseFields($course, $newCourse); 
-        $course->save();
-        return true;
+        if(is_null(Course::where('sigaa_id', $newCourse['sigaa_id'])->first())){
+            $course = Course::fillCourseFields($course, $newCourse); 
+            $course->save();
+            return true;
+        } 
+        return false;
     }
 
     public function editCourse($oldSigaaId, $course){
         $editableCourse = Course::checkIfCourseAlreadyExist($oldSigaaId);
+        if(is_null($editableCourse)){
+            return "error";
+        }
         $editableCourse = Course::fillCourseFields($editableCourse, $course);
         $editableCourse->save();
+        return true;
     }
 
     public function deleteCourse($sigaaId){
         $course = Course::checkIfCourseAlreadyExist($sigaaId);
+        if(is_null($course)){
+            return "error";
+        }
         $course->delete();
+        return true;
     }
 
     public function findCourseByName($courseName){
-        return Course::where('name', $courseName)->first();
+        $course = new Course();
+        $course = Course::where('name', $courseName)->first();
+        if(is_null($course)){
+            return "error";
+        }
+        return $course;
     }
 
     public function findBySigaaID($sigaaId){
@@ -46,7 +62,12 @@ class Course extends Model
     }
 
     protected function checkIfCourseAlreadyExist($sigaaId){
-        return Course::find($sigaaId);
+        $course = new Course();
+        $course = Course::find($sigaaId);
+        if(is_null($course)){
+            return "error";
+        }
+        return $course;
     }
 
     protected function fillCourseFields($course, $courseArray){
